extract favorites section into its own file
